Extracts interface from PushbackSideInputDoFnRunner
